third_party/WebKit/LayoutTests/webaudio/Panner/panner-automation-basic-expected.txt

Test Basic PannerNode with Automation Position Properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Stereo panner.positionX.value frame [0, 128) channel 0 contains only the constant 6.123234262925839e-17.
PASS Stereo panner.positionX.value frame [0, 128) channel 1 contains only the constant 3.
PASS Stereo panner.positionX.value frame [128, 256) channel 0 contains only the constant 6.123233968888981e-21.
PASS Stereo panner.positionX.value frame [128, 256) channel 0 contains only the constant 0.00029999998514540493.
PASS Output at frame 128 channel 0 is not equal to 6.123234262925839e-17.
PASS Output at frame 128 channel 1 is not equal to 3.
PASS Directly setting Stereo panner.positionX.value worked.

PASS Mono panner.positionX.value frame [0, 128) channel 0 contains only the constant 6.123234262925839e-17.
PASS Mono panner.positionX.value frame [0, 128) channel 1 contains only the constant 1.
PASS Mono panner.positionX.value frame [128, 256) channel 0 contains only the constant 6.123233968888981e-21.
PASS Mono panner.positionX.value frame [128, 256) channel 0 contains only the constant 0.00009999999747378752.
PASS Output at frame 128 channel 0 is not equal to 6.123234262925839e-17.
PASS Output at frame 128 channel 1 is not equal to 1.
PASS Directly setting Mono panner.positionX.value worked.

PASS Stereo panner.positionY.value frame [0, 128) channel 0 contains only the constant 6.123234262925839e-17.
PASS Stereo panner.positionY.value frame [0, 128) channel 1 contains only the constant 3.
PASS Stereo panner.positionY.value frame [128, 256) channel 0 contains only the constant 6.123233968888981e-21.
PASS Stereo panner.positionY.value frame [128, 256) channel 0 contains only the constant 0.00029999998514540493.
PASS Output at frame 128 channel 0 is not equal to 6.123234262925839e-17.
PASS Output at frame 128 channel 1 is not equal to 3.
PASS Directly setting Stereo panner.positionY.value worked.

PASS Mono panner.positionY.value frame [0, 128) channel 0 contains only the constant 6.123234262925839e-17.
PASS Mono panner.positionY.value frame [0, 128) channel 1 contains only the constant 1.
PASS Mono panner.positionY.value frame [128, 256) channel 0 contains only the constant 6.123233968888981e-21.
PASS Mono panner.positionY.value frame [128, 256) channel 0 contains only the constant 0.00009999999747378752.
PASS Output at frame 128 channel 0 is not equal to 6.123234262925839e-17.
PASS Output at frame 128 channel 1 is not equal to 1.
PASS Directly setting Mono panner.positionY.value worked.

PASS Stereo panner.positionZ.value frame [0, 128) channel 0 contains only the constant 1.
PASS Stereo panner.positionZ.value frame [0, 128) channel 1 contains only the constant 2.
PASS Stereo panner.positionZ.value frame [128, 256) channel 0 contains only the constant 0.00009999999747378752.
PASS Stereo panner.positionZ.value frame [128, 256) channel 0 contains only the constant 0.00019999999494757503.
PASS Output at frame 128 channel 0 is not equal to 1.
PASS Output at frame 128 channel 1 is not equal to 2.
PASS Directly setting Stereo panner.positionZ.value worked.

PASS Mono panner.positionZ.value frame [0, 128) channel 0 contains only the constant 0.7071067690849304.
PASS Mono panner.positionZ.value frame [0, 128) channel 1 contains only the constant 0.7071067690849304.
PASS Mono panner.positionZ.value frame [128, 256) channel 0 contains only the constant 0.00007071067375363782.
PASS Mono panner.positionZ.value frame [128, 256) channel 0 contains only the constant 0.00007071067375363782.
PASS Output at frame 128 channel 0 is not equal to 0.7071067690849304.
PASS Output at frame 128 channel 1 is not equal to 0.7071067690849304.
PASS Directly setting Mono panner.positionZ.value worked.

PASS Stereo listener.positionX.value frame [0, 128) channel 0 contains only the constant 1.
PASS Stereo listener.positionX.value frame [0, 128) channel 1 contains only the constant 2.
PASS Stereo listener.positionX.value frame [128, 256) channel 0 contains only the constant 0.00030002999119460583.
PASS Stereo listener.positionX.value frame [128, 256) channel 0 contains only the constant 0.
PASS Output at frame 128 channel 0 is not equal to 1.
PASS Output at frame 128 channel 1 is not equal to 2.
PASS Directly setting Stereo listener.positionX.value worked.

PASS Mono listener.positionX.value frame [0, 128) channel 0 contains only the constant 0.7071067690849304.
PASS Mono listener.positionX.value frame [0, 128) channel 1 contains only the constant 0.7071067690849304.
PASS Mono listener.positionX.value frame [128, 256) channel 0 contains only the constant 0.00010001000191550702.
PASS Mono listener.positionX.value frame [128, 256) channel 0 contains only the constant 0.
PASS Output at frame 128 channel 0 is not equal to 0.7071067690849304.
PASS Output at frame 128 channel 1 is not equal to 0.7071067690849304.
PASS Directly setting Mono listener.positionX.value worked.

PASS Stereo listener.positionY.value frame [0, 128) channel 0 contains only the constant 0.40824827551841736.
PASS Stereo listener.positionY.value frame [0, 128) channel 1 contains only the constant 1.5629488229751587.
PASS Stereo listener.positionY.value frame [128, 256) channel 0 contains only the constant 0.0000707106664776802.
PASS Stereo listener.positionY.value frame [128, 256) channel 0 contains only the constant 0.00027071067597717047.
PASS Output at frame 128 channel 0 is not equal to 0.40824827551841736.
PASS Output at frame 128 channel 1 is not equal to 1.5629488229751587.
PASS Directly setting Stereo listener.positionY.value worked.

PASS Mono listener.positionY.value frame [0, 128) channel 0 contains only the constant 0.22094237804412842.
PASS Mono listener.positionY.value frame [0, 128) channel 1 contains only the constant 0.5334020853042603.
PASS Mono listener.positionY.value frame [128, 256) channel 0 contains only the constant 0.00003826833926723339.
PASS Mono listener.positionY.value frame [128, 256) channel 0 contains only the constant 0.00009238795610144734.
PASS Output at frame 128 channel 0 is not equal to 0.22094237804412842.
PASS Output at frame 128 channel 1 is not equal to 0.5334020853042603.
PASS Directly setting Mono listener.positionY.value worked.

PASS Stereo listener.positionZ.value frame [0, 128) channel 0 contains only the constant 6.123234262925839e-17.
PASS Stereo listener.positionZ.value frame [0, 128) channel 1 contains only the constant 3.
PASS Stereo listener.positionZ.value frame [128, 256) channel 0 contains only the constant 0.00010001000191550702.
PASS Stereo listener.positionZ.value frame [128, 256) channel 0 contains only the constant 0.00020003000099677593.
PASS Output at frame 128 channel 0 is not equal to 6.123234262925839e-17.
PASS Output at frame 128 channel 1 is not equal to 3.
PASS Directly setting Stereo listener.positionZ.value worked.

PASS Mono listener.positionZ.value frame [0, 128) channel 0 contains only the constant 6.123234262925839e-17.
PASS Mono listener.positionZ.value frame [0, 128) channel 1 contains only the constant 1.
PASS Mono listener.positionZ.value frame [128, 256) channel 0 contains only the constant 0.00007071421714499593.
PASS Mono listener.positionZ.value frame [128, 256) channel 0 contains only the constant 0.0000707212820998393.
PASS Output at frame 128 channel 0 is not equal to 6.123234262925839e-17.
PASS Output at frame 128 channel 1 is not equal to 1.
PASS Directly setting Mono listener.positionZ.value worked.

PASS setPosition.value frame [0, 128) channel 0 contains only the constant 0.40824827551841736.
PASS setPosition.value frame [0, 128) channel 1 contains only the constant 1.5629488229751587.
PASS setPosition.value frame [128, 256) channel 0 contains only the constant 0.000124290119856596.
PASS setPosition.value frame [128, 256) channel 0 contains only the constant 0.0002424988488201052.
PASS Output at frame 128 channel 0 is not equal to 0.40824827551841736.
PASS Output at frame 128 channel 1 is not equal to 1.5629488229751587.
PASS Directly setting setPosition.value worked.

PASS panner.orientation{XYZ}.value frame [0, 128) channel 0 contains only the constant 0.5005000233650208.
PASS panner.orientation{XYZ}.value frame [0, 128) channel 1 contains only the constant 1.0010000467300415.
PASS panner.orientation{XYZ}.value frame [128, 256) channel 0 contains only the constant 0.07899462431669235.
PASS panner.orientation{XYZ}.value frame [128, 256) channel 0 contains only the constant 0.1579892486333847.
PASS Output at frame 128 channel 0 is not equal to 0.5005000233650208.
PASS Output at frame 128 channel 1 is not equal to 1.0010000467300415.
PASS Directly setting panner.orientation{XYZ}.value worked.

PASS listener.forward{XYZ}.value frame [0, 128) channel 0 contains only the constant 0.5005000233650208.
PASS listener.forward{XYZ}.value frame [0, 128) channel 1 contains only the constant 1.0010000467300415.
PASS listener.forward{XYZ}.value frame [128, 256) channel 0 contains only the constant 0.5133963823318481.
PASS listener.forward{XYZ}.value frame [128, 256) channel 0 contains only the constant 1.0009169578552246.
PASS Output at frame 128 channel 0 is not equal to 0.5005000233650208.
PASS Output at frame 128 channel 1 is not equal to 1.0010000467300415.
PASS Directly setting listener.forward{XYZ}.value worked.

PASS listener.up{XYZ}.value frame [0, 128) channel 0 contains only the constant 0.12537498772144318.
PASS listener.up{XYZ}.value frame [0, 128) channel 1 contains only the constant 0.47998905181884766.
PASS listener.up{XYZ}.value frame [128, 256) channel 0 contains only the constant 0.1447705775499344.
PASS listener.up{XYZ}.value frame [128, 256) channel 0 contains only the constant 0.4569823145866394.
PASS Output at frame 128 channel 0 is not equal to 0.12537498772144318.
PASS Output at frame 128 channel 1 is not equal to 0.47998905181884766.
PASS Directly setting listener.up{XYZ}.value worked.

PASS successfullyParsed is true

TEST COMPLETE
